Exercise02
You can see all the flags on one page in turn. I made the last flag the way I wanted it to feel. I painted the background to make my favourite colour, red, feel like a gradient. Red reminds me of a heart, so I inserted the image as a point :)